Refactored code into separate directories
