+++ src/styles/NewNoteSection.css
