+++ src/styles/Note.css
